adds header Authorization = Bearer {{bearerToken}}
Before issuing GET request for the resource group you need to add a header for Authorization to use {{bearerToken}} value set in the test section above. You may already have it set but this article does not call for it and without setting that value it isn't working. You may want to take another screenshot and add it.